GEOD-379 - FormArrayNames simple version refactor
